THE FINAL COMMIT
After much trial and error, I made the ultimate addition to my magpie
file: it can deal with compound sentences.  If you give it a
multi-clause answer, it will decide which one it thinks is the most
important and respond to it only.  I also cleaned up the “how” question
answers.  It now answers with pluralized verbs and correct pronouns
most of the time.