The Last Contraction
So, obviously my last commit summary was a bit premature.  This time, I
added an extra method that handles the ’s contraction.  The tricky
thing is, sometimes it is a contraction, and sometimes it is possessive
tense.  What I did was make it so that if and only if there is no verb
in the clause, it will expand ’s to is.